pyproject.toml
[project]
name = "f8848"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.135.1",
    "jinja2>=3.1.6",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "tushare>=1.4.24",
    "uvicorn>=0.41.0",
    "python-multipart",
]
